+++ rotkehlchen/chain/arbitrum_one/modules/thegraph/balances.py
@@ -6,6 +6,7 @@ from rotkehlchen.chain.evm.decoding.thegraph.balances import ThegraphCommonBalan
 from rotkehlchen.constants.assets import A_GRT_ARB
 
 if TYPE_CHECKING:
+    from rotkehlchen.chain.arbitrum_one.decoding.decoder import ArbitrumOneTransactionDecoder
     from rotkehlchen.chain.arbitrum_one.node_inquirer import ArbitrumOneInquirer
     from rotkehlchen.db.dbhandler import DBHandler
 
@@ -16,10 +17,12 @@ class ThegraphBalances(ThegraphCommonBalances):
             self,
             database: 'DBHandler',
             evm_inquirer: 'ArbitrumOneInquirer',
+            tx_decoder: 'ArbitrumOneTransactionDecoder',
     ) -> None:
         super().__init__(
             database=database,
             evm_inquirer=evm_inquirer,
+            tx_decoder=tx_decoder,
             native_asset=A_GRT_ARB,
             staking_contract=CONTRACT_STAKING,
         )