+++ rotkehlchen/chain/ethereum/modules/blur/balances.py
@@ -19,6 +19,7 @@ from rotkehlchen.inquirer import Inquirer
 from rotkehlchen.logging import RotkehlchenLogsAdapter
 
 if TYPE_CHECKING:
+    from rotkehlchen.chain.ethereum.decoding.decoder import EthereumTransactionDecoder
     from rotkehlchen.chain.evm.node_inquirer import EvmNodeInquirer
     from rotkehlchen.db.dbhandler import DBHandler
     from rotkehlchen.types import ChecksumEvmAddress
@@ -32,10 +33,12 @@ class BlurBalances(ProtocolWithBalance):
             self,
             database: 'DBHandler',
             evm_inquirer: 'EvmNodeInquirer',
+            tx_decoder: 'EthereumTransactionDecoder',
     ):
         super().__init__(
             database=database,
             evm_inquirer=evm_inquirer,
+            tx_decoder=tx_decoder,
             counterparty=CPT_BLUR,
             deposit_event_types={(HistoryEventType.STAKING, HistoryEventSubType.DEPOSIT_ASSET)},
         )